packages/ancol-common/pyproject.toml
[project]
name = "ancol-common"
version = "0.1.0"
description = "Shared utilities for Ancol MoM Compliance System"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "google-genai>=1.0",
    "google-cloud-pubsub>=2.0",
    "google-cloud-storage>=2.0",
    "google-cloud-bigquery>=3.0",
    "google-cloud-secret-manager>=2.0",
    "fastapi>=0.115",
    "httpx>=0.27",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ancol_common"]
